+++ app/(tabs)/builder/build-step-1.tsx
@@ -5,7 +5,7 @@ import { ThemedText } from '@/components/ThemedText';
 import { ThemedView } from '@/components/ThemedView';
 import { IconSymbol } from '@/components/ui/IconSymbol';
 
-export default function BuildStep2() {
+export default function BuildStep1() {
   return (
     <ParallaxScrollView
       headerBackgroundColor={{ light: '#D0D0D0', dark: '#353636' }}
@@ -18,9 +18,9 @@ export default function BuildStep2() {
         />
       }>
       <ThemedView style={styles.titleContainer}>
-        <ThemedText type="title">Build Step 2</ThemedText>
+        <ThemedText type="title">Build Step 1</ThemedText>
       </ThemedView>
-      <ThemedText>This is where we're looking to add Archetype.</ThemedText>
+      <ThemedText>This is where we&#39;re looking to add Archetype.</ThemedText>
     </ParallaxScrollView>
   );
 }
