+++ components/ui/IconSymbol.tsx
@@ -19,6 +19,8 @@ const MAPPING = {
   'chevron.left.forwardslash.chevron.right': 'code',
   'chevron.right': 'chevron-right',
   'person.fill': 'person',
+  'person.fill.badge.plus': 'person-add',
+  'plus.app.fill': 'add-box',
 } as IconMapping;
 
 /**
